Move SecureChannel to transport folder
