+++ src/lib/core/tests/Makefile.am
@@ -43,7 +43,6 @@ lib_LIBRARIES                                         = \
 libCoreTests_a_SOURCES                                = \
     TestCHIPErrorStr.cpp                                \
     TestCHIPTLV.cpp                                     \
-    TestCHIPSecureChannel.cpp                           \
     $(NULL)
 
 libCoreTests_adir                                     = $(includedir)/core
@@ -88,7 +87,6 @@ COMMON_LDADD                                          =  \
 check_PROGRAMS                                        = \
     TestCHIPErrorStr                                    \
     TestCHIPTLV                                         \
-    TestCHIPSecureChannel                               \
     $(NULL)
 
 # Test applications and scripts that should be built and run when the
@@ -111,9 +109,6 @@ TestCHIPErrorStr_LDADD                                = $(COMMON_LDADD)
 
 TestCHIPTLV_SOURCES                                   = TestCHIPTLVDriver.cpp
 TestCHIPTLV_LDADD                                     = $(COMMON_LDADD)
-
-TestCHIPSecureChannel_SOURCES                         = TestCHIPSecureChannelDriver.cpp
-TestCHIPSecureChannel_LDADD                           = $(COMMON_LDADD)
 
 #
 # Foreign make dependencies
