+++ src/transport/TransportLayer.am
@@ -27,9 +27,11 @@
 CHIP_BUILD_TRANSPORT_LAYER_SOURCE_FILES                  = \
     @top_builddir@/src/transport/MessageHeader.cpp         \
     @top_builddir@/src/transport/UdpTransport.cpp          \
+    @top_builddir@/src/transport/CHIPSecureChannel.cpp     \
     $(NULL)
 
 CHIP_BUILD_TRANSPORT_LAYER_HEADER_FILES  = \
     MessageHeader.h                        \
     UdpTransport.h                         \
+    CHIPSecureChannel.h                    \
     $(NULL)
