+++ src/transport/tests/Makefile.am
@@ -43,6 +43,7 @@ lib_LIBRARIES                                         = \
 libTransportLayerTests_a_SOURCES                      = \
     TestMessageHeader.cpp                               \
     TestUdpTransport.cpp                                \
+    TestCHIPSecureChannel.cpp                           \
     $(NULL)
 
 libTransportLayerTests_adir                           = $(includedir)/transport
@@ -86,6 +87,7 @@ COMMON_LDADD                                          = \
 check_PROGRAMS             = \
     TestMessageHeader        \
     TestUdpTransport         \
+    TestCHIPSecureChannel    \
     $(NULL)
 
 # Test applications and scripts that should be built and run when the
@@ -108,6 +110,9 @@ TestMessageHeader_LDADD       = $(COMMON_LDADD)
 
 TestUdpTransport_SOURCES     = TestUdpTransportDriver.cpp
 TestUdpTransport_LDADD       = $(COMMON_LDADD)
+
+TestCHIPSecureChannel_SOURCES                         = TestCHIPSecureChannelDriver.cpp
+TestCHIPSecureChannel_LDADD                           = $(COMMON_LDADD)
 
 #
 # Foreign make dependencies
